+++ test/reviewer_test.rb
@@ -356,6 +356,18 @@ module Reviewer
       refute_match(/"version"/, output)
     end
 
+    def test_unknown_option_is_a_usage_error
+      out, _err = with_argv('--list') do
+        capture_subprocess_io do
+          error = assert_raises(SystemExit) { Reviewer.review }
+          assert_equal Session::USAGE_ERROR, error.status
+        end
+      end
+
+      assert_match(/unknown option `--list'/, out)
+      assert_match(/rvw --help/, out)
+    end
+
     private
 
     def with_argv(*args)